feat(fix): update al files related to Venta Class, upload enums directory and EstadoVenta.java and MetodoPago.java files
